requirements.txt
APScheduler==3.11.0
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.1.3
django-apscheduler==0.7.0
django-cors-headers==4.6.0
django-environ==0.11.2
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0